move archetypeNode to separate file
